MINOR: config: Deprecate and ignore tune.chksize global option
This option is now ignored because I/O check buffers are now allocated using the
buffer pool. Thus, it is marked as deprecated in the documentation and ignored
during the configuration parsing. The field is also removed from the global
structure.

Because this option is ignored since a recent fix, backported as fare as 2.2,
this patch should be backported too. Especially because it updates the
documentation.

(cherry picked from commit bb9fb8b7f832d3a2cb62f5b4046390ae7943fc16)
Signed-off-by: Willy Tarreau <w@1wt.eu>