CLEANUP: lua: Remove declaration of an inexistant function
action_http_after_res_custom() is declared but never defined. The declaration
was introduced by the commit 86f8f6f78 ("MINOR: actions: Export actions lookup
functions"). But the definition does not exist because the HTTP after-response
ruleset only exist since the 2.2.

This patch must be backported where the above patch is, as far as 1.8. There is
no upstream commit ID for this patch.